Kluge, ALL devices in pushbullet
This is a total kluge, it removes the device ID attribute from being using in pushbullet which, will send to all devices.  I know know much about perl and was unable to craft an if/then/else statement that would use the device ID if it wasn't set to "all".

I kept getting errors telling me that $responce was undefined, which was extremely frustrating but likely entirely my fault for fussing with a script written in a language I don't know.

The new pushbullet API says if you don't specify a device ID (or anywhere else to send a push) it just sends it out to all devices on a pushbullet account.